requirements.txt
streamlit
pandas
openpyxl
matplotlib
img2pdf
pillow
